Added bindings to libscran's new downsampling functions.

+++ src/build_nn_index.cpp
@@ -9,9 +9,9 @@ SEXP build_nn_index(Rcpp::NumericMatrix data, bool approximate) {
     size_t nr = data.nrow(), nc = data.ncol();
     auto ptr = static_cast<const double*>(data.begin());
     if (approximate) {
-        return KnncollePtr(new knncolle::AnnoyEuclidean<int, float>(nr, nc, ptr));
+        return KnncollePtr(new knncolle::AnnoyEuclidean<int, double>(nr, nc, ptr));
     } else {
-        return KnncollePtr(new knncolle::VpTreeEuclidean<int, float>(nr, nc, ptr));
+        return KnncollePtr(new knncolle::VpTreeEuclidean<int, double>(nr, nc, ptr));
     }
 }
