+++ src/knncolle.h
@@ -4,7 +4,7 @@
 #include "knncolle/utils/Base.hpp"
 #include "Rcpp.h"
 
-typedef Rcpp::XPtr<knncolle::Base<int, float> > KnncollePtr;
+typedef Rcpp::XPtr<knncolle::Base<int, double> > KnncollePtr;
 
 #include <vector>
